requirements.txt
fastapi
uvicorn
python-multipart
pymongo
motor
python-dotenv
pillow
passlib
bcrypt